User create

+++ src/main/java/kg/easyit/onlineshop/model/entity/Product.java
@@ -25,7 +25,7 @@ public class Product extends AbstractPersistable<Long> {
     BigDecimal price;
 
     @Column(name = "units_in_stock", nullable = false)
-    Long unitsInStock;
+    Integer unitsInStock;
 
     @UpdateTimestamp
     @Column(name = "date_updated", nullable = false)
